requirements.txt
anthropic>=0.40.0
google-cloud-bigquery>=3.0.0
